tests/wpt/web-platform-tests/encoding-detection/ar-ISO-8859-6-late.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>ar ISO-8859-6 late</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "ISO-8859-6", 'Expected ISO-8859-6');
    }, "Check detection result");
    w.close();
    done();
};
var w = window.open("support/ar-ISO-8859-6-late.sub.html");
</script>
